requirements.txt
streamlit>=1.14.0
sqlalchemy>=1.4.0
openai>=0.27.0
tiktoken>=0.4.0
python-dotenv>=0.19.0
pyyaml>=5.4.1
langchain>=0.0.200
neo4j>=4.4.0
chromadb>=0.3.22
pytest>=7.1.0
numpy>=1.21.0
neo4j
sentence-transformers>=2.2.2
ragxplorer>=0.1.0 # Add ragxplorer
